FlaskApi/src/requirements.txt
astroid==1.6.3
click==6.7
Flask==0.12.2
flask-heroku==0.1.9
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.4
PyJWT==1.4.2
pylint==1.8.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
wrapt==1.10.11
